Update README.md
This project’s objectives:

Rigor, use of C, use
of basic algorithms, information research, data mining etc.
As a graphic design project, Wolf3D will enable you to solidify your skills in this area:
windows, colors, events, fill shapes etc.


To conclude Wolf3D is a remarkable playground to explore the playful practical applications
of mathematics without having to understand the specifics. John Carmack is
undoubtedly a genius. He was one of the first programmers to think about these applications
and remains famous for them today. With the help of the numerous documents
available on the internet, you will use mathematics as a tool to create elegant and efficient
algorithms.